Applied renamings to imports
