+++ examples/example_5.py
@@ -5,9 +5,9 @@ sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
 
 
 # Importing game module
-from game.interface import main
-from game.agents.cmd_agent import CMDAgent
-from game.agents.personality_agent import PersonalityAgent
+from src.api import api
+from src.agents.cmd import CMDAgent
+from src.agents.personality import PersonalityAgent
 
 
 # Importing other modules
