+++ src/agents/personality.py
@@ -1,6 +1,6 @@
-from .base_agent import BaseAgent
+from .base import BaseAgent
 import random
-from ..core import constants
+from ..engine import constants
 from ..utils import map_range
 from rich.traceback import install
 install()
